Improve flow types
