+++ example/components/Heart.js
@@ -1,0 +1,4 @@
+// @flow
+import React from 'react';
+
+
