+++ lib/index.js
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+export {default as Transform} from 'art/core/transform';
+export {default as Path} from './ARTSerializablePath';
+export {default as Surface} from './Surface';
+export {default as Group} from './Group';
+export {default as Shape} from './Shape';
+export {default as Text} from './Text';
+export {default as ClippingRectangle} from './ClippingRectangle';
+export {default as LinearGradient} from './LinearGradient';
+export {default as RadialGradient} from './RadialGradient';
+export {default as Pattern} from './Pattern';
