+++ lib/nativeComponents.js
@@ -1,0 +1,17 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+import {requireNativeComponent} from 'react-native';
+
+const NativeSurfaceView = requireNativeComponent('ARTSurfaceView');
+const NativeGroup = requireNativeComponent('ARTGroup');
+const NativeShape = requireNativeComponent('ARTShape');
+const NativeText = requireNativeComponent('ARTText');
+
+export {NativeSurfaceView, NativeGroup, NativeShape, NativeText};
